+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ releases, in reverse chronological order.
   * ``ReceiptPDFDisplayView`` has been dropped in favour of the above.
   * ``ReceiptPDFDownloadView`` prompts users to download a receipt's PDF. The
     PDF's file name is now customizable by overriding ``get_download_name``.
+* Allow filtering receipts [in the admin] by type.
 
 4.1.7
 -----